gnupg20: Import stripped-down 2.0.31.
Updated and modified version of joyent/gnupg20, stripping out everything
except for the gpg2 binary, as gpg-agent in this version no longer
builds against libassuan 3.x in pkgsrc trunk.

Hosts will need to run gpg-agent from an older version of this package,
and then use this one for signing packages.